requirements.txt
Django==5.0
django-allauth==0.60.1
django-cors-headers==4.3.1
django-pwa==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf_spectacular==0.27.1
gunicorn==21.2.0
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
whitenoise==6.6.0
